+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to the "robotcode" extension will be documented in this file
   - In an expression like `${A+'${B+"${F}"}'+'${D}'} ${C}` for example, every single 'inner' variable will be recognized, you can hover over it, it can be found as reference, you can go to the definition, ...
   - The debugger shows all variables as inline values and when hovering, it shows the values of the variables.
   - Only the name of the variable is used, not the surrounding ${}
+- Support importing variable files as module for RobotFramework 5
 
 ##  0.5.5